Add Model Classes, update constant file
